restructuring and organization of the multicamdialog
