+++ client/dive-common/components/ImportMultiCamChooseAnnotation.vue
@@ -1,38 +1,5 @@
 <script lang="ts">
-import { defineComponent } from 'vue';
+import ImportMultiCamChooseAnnotation from 'dive-common/components/ImportMultiCamDialog/ImportMultiCamChooseAnnotation.vue';
 
-export default defineComponent({
-  props: {
-    trackFile: String,
-    cameraName: String,
-  },
-});
+export default ImportMultiCamChooseAnnotation;
 </script>
-
-<template>
-  <v-row
-    class="align-center"
-    no-gutters
-  >
-    <v-text-field
-      label="Annotation File"
-      :placeholder="`Annotation File (${cameraName})`"
-      outlined
-      dense
-      hide-details
-      clearable
-      :value="trackFile"
-      class="mr-3"
-      @click:clear="$emit('clear')"
-    />
-    <v-btn
-      color="primary"
-      @click="$emit('open')"
-    >
-      Open
-      <v-icon class="ml-2">
-        mdi-file-table
-      </v-icon>
-    </v-btn>
-  </v-row>
-</template>
